updates for fuzz
