+++ src/pos/wallet/kernelrecord.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <pos/kernelrecord.h>
+#include <pos/wallet/kernelrecord.h>
 
 #include <pos/kernel.h>
 #include <base58.h>
