+++ src/pos/wallet/signer.h
@@ -2,8 +2,8 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#ifndef BITCOIN_POS_SIGNER_H
-#define BITCOIN_POS_SIGNER_H
+#ifndef BITCOIN_POS_WALLET_SIGNER_H
+#define BITCOIN_POS_WALLET_SIGNER_H
 
 #include <primitives/block.h>
 #include <primitives/transaction.h>
@@ -13,6 +13,7 @@ class CBlock;
 class CWallet;
 class LegacyScriptPubKeyMan;
 
-bool CheckBlockSignature(const CBlock& block);
 
-#endif // BITCOIN_POS_SIGNER_H
+bool SignBlock(CBlock& block, const CWallet& keystore);
+
+#endif // BITCOIN_POS_WALLET_SIGNER_H
