+++ src/pos/wallet/stake.h
@@ -3,17 +3,18 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#ifndef POS_STAKE_H
-#define POS_STAKE_H
+#ifndef BITCOIN_POS_WALLET_STAKE_H
+#define BITCOIN_POS_WALLET_STAKE_H
 
 #include <consensus/params.h>
 #include <wallet/wallet.h>
 #include <validation.h>
 
 class CChainState;
+class CWallet;
 
-bool GetStakeWeight(const CWallet* pwallet, uint64_t& nAverageWeight, uint64_t& nTotalWeight, const Consensus::Params& consensusParams);
+bool GetStakeWeight(const CWallet* pwallet, CChainState* chainstate, uint64_t& nAverageWeight, uint64_t& nTotalWeight, const Consensus::Params& consensusParams);
 bool CreateCoinStake(const CWallet* pwallet, CChainState* chainstate, unsigned int nBits, int64_t nSearchInterval, CMutableTransaction& txNew, const Consensus::Params& consensusParams, const CAmount& nFees);
 
-#endif // POS_STAKE_H
+#endif // BITCOIN_POS_WALLET_STAKE_H
