+++ src/test/fuzz/bech32.cpp
@@ -32,11 +32,11 @@ FUZZ_TARGET(bech32)
     if (input.size() + 3 + 6 <= 90) {
         // If it's possible to encode input in Bech32(m) without exceeding the 90-character limit:
         for (auto encoding : {bech32::Encoding::BECH32, bech32::Encoding::BECH32M}) {
-            const std::string encoded = bech32::Encode(encoding, "bc", input);
+            const std::string encoded = bech32::Encode(encoding, "indx", input);
             assert(!encoded.empty());
             const auto r2 = bech32::Decode(encoded);
             assert(r2.encoding == encoding);
-            assert(r2.hrp == "bc");
+            assert(r2.hrp == "indx");
             assert(r2.data == input);
         }
     }
