+++ src/test/fuzz/pow.cpp
@@ -60,10 +60,11 @@ FUZZ_TARGET_INIT(pow, initialize_pow)
         }
         {
             (void)GetBlockProof(current_block);
-            (void)CalculateNextWorkRequired(arith_uint256(20000), 60, 60, false, consensus_params);
-            if (current_block.nHeight != std::numeric_limits<int>::max() && current_block.nHeight - (consensus_params.DifficultyAdjustmentInterval() - 1) >= 0) {
+            // Diff at every block
+            //(void)CalculateNextWorkRequired(arith_uint256(20000), 60, 60, false, consensus_params);
+            //if (current_block.nHeight != std::numeric_limits<int>::max() && current_block.nHeight - (consensus_params.DifficultyAdjustmentInterval() - 1) >= 0) {
                 (void)GetNextWorkRequired(&current_block, &(*block_header), consensus_params);
-            }
+            //}
         }
         {
             const CBlockIndex* to = &PickValue(fuzzed_data_provider, blocks);
