+++ src/test/fuzz/process_messages.cpp
@@ -37,7 +37,7 @@ FUZZ_TARGET_INIT(process_messages, initialize_process_messages)
 
     ConnmanTestMsg& connman = *static_cast<ConnmanTestMsg*>(g_setup->m_node.connman.get());
     TestChainState& chainstate = *static_cast<TestChainState*>(&g_setup->m_node.chainman->ActiveChainstate());
-    SetMockTime(1610000000); // any time to successfully reset ibd
+    SetMockTime(1654238500); // any time to successfully reset ibd
     chainstate.ResetIbd();
 
     std::vector<CNode*> peers;
