+++ src/wallet/wallet.cpp
@@ -1717,6 +1717,7 @@ void CWallet::ReacceptWalletTransactions()
 
 void CWallet::AbandonOrphanedCoinstakes()
 {
+    LOCK(cs_wallet);
     for (std::pair<const uint256, CWalletTx>& item : mapWallet) {
         const uint256& wtxid = item.first;
         CWalletTx& wtx = item.second;
